Move from_database functions to separate file
